jaato-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jaato-server"
version = "0.2.6"
description = "Jaato Server - LLM tool orchestration runtime and daemon"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "apanoia"}
]
keywords = ["llm", "ai", "tools", "orchestration", "vertex-ai", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "jaato-sdk",
    "click",
    "requests",
    "google-genai",
    "mcp[cli]",
    "google-api-core",
    "google-auth",
    "certifi",
    "python-dotenv",
    "pyreadline3",
    "reportlab",
    "ddgs>=7.0.0",
    "pyyaml",
    "azure-ai-inference>=1.0.0b7",
    "websockets>=12.0",
    "anthropic>=0.40.0",
    "pybars3>=0.9.7",
    "httpx>=0.27.0",
    "trafilatura>=1.6.0",
    "html2text>=2024.2.26",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "ipython-pygments-lexers>=1.0.0",
    "Pillow>=10.0.0",
    "pexpect>=4.8.0; sys_platform != 'win32'",
    "wexpect>=4.0.0; sys_platform == 'win32'",
    "pywin32>=220; sys_platform == 'win32'",  # wexpect undeclared dep
    "psutil>=5.0.0; sys_platform == 'win32'",  # wexpect undeclared dep
    "setuptools; sys_platform == 'win32'",  # wexpect requires pkg_resources
    "pyte>=0.8.0",
    "watchdog>=4.0.0",
    "rich>=13.0.0",  # Used by shared plugins for formatting
    "ast-grep-py>=0.30.0",  # AST-based code search
    "pymupdf4llm>=0.2.0",  # PDF to markdown conversion for web_fetch
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
kaggle = [
    "kaggle>=1.5.0",
]
kerberos = [
    "pyspnego[kerberos]>=0.10.0",  # Kerberos/SPNEGO proxy authentication (includes gssapi on Linux/macOS)
]
all = [
    "jaato-server[dev,kaggle,kerberos]",
]

[project.scripts]
jaato-server = "server.__main__:main"

# Entry points for plugin discovery
# External packages can register plugins by adding to this group
[project.entry-points."jaato.plugins"]
artifact_tracker = "shared.plugins.artifact_tracker:create_plugin"
background = "shared.plugins.background:create_plugin"
calculator = "shared.plugins.calculator:create_plugin"
cli = "shared.plugins.cli:create_plugin"
environment = "shared.plugins.environment:create_plugin"
mcp = "shared.plugins.mcp:create_plugin"
lsp = "shared.plugins.lsp:create_plugin"
memory = "shared.plugins.memory:create_plugin"
multimodal = "shared.plugins.multimodal:create_plugin"
permission = "shared.plugins.permission:create_plugin"
references = "shared.plugins.references:create_plugin"
template = "shared.plugins.template:create_plugin"
todo = "shared.plugins.todo:create_plugin"
anthropic_auth = "shared.plugins.anthropic_auth:create_plugin"
ast_search = "shared.plugins.ast_search:create_plugin"

# Entry points for GC strategy plugins
# GC plugins implement context window garbage collection strategies
[project.entry-points."jaato.gc_plugins"]
gc_truncate = "shared.plugins.gc_truncate:create_plugin"
gc_summarize = "shared.plugins.gc_summarize:create_plugin"
gc_hybrid = "shared.plugins.gc_hybrid:create_plugin"
gc_budget = "shared.plugins.gc_budget:create_plugin"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "server*", "jaato*"]
